+++ NetEvent/Server/Modules/Roles/Endpoints/GetRoles/GetRolesRequest.cs
@@ -1,6 +1,6 @@
 ﻿using MediatR;
 
-namespace NetEvent.Server.Modules.Roles.Endpoints
+namespace NetEvent.Server.Modules.Roles.Endpoints.GetRoles
 {
     public class GetRolesRequest : IRequest<GetRolesResponse>
     {
