+++ NetEvent/Server/Modules/Roles/Endpoints/GetRoles/GetRolesResponse.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using NetEvent.Shared.Dto;
 
-namespace NetEvent.Server.Modules.Roles.Endpoints
+namespace NetEvent.Server.Modules.Roles.Endpoints.GetRoles
 {
     public class GetRolesResponse : ResponseBase<IReadOnlyCollection<RoleDto>>
     {
